Time: 0 ms (100%), Space: 8.7 MB (20.55%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.cpp
@@ -3,7 +3,7 @@ public:
     int reverse(int x) {
         int ans=0;
         while(x!=0){
-            long long int digit=x%10;
+            int digit=x%10;
             if(ans> INT_MAX/10 || ans < INT_MIN/10) return 0;
             ans=(ans*10)+digit;
             x/=10;
